# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Admin API",
        "description": "API Documentation for the Laravel MongoDB MVC backend.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "List lessons",
                "operationId": "listLessons",
                "responses": {
                    "200": {
                        "description": "Lesson collection.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Lesson"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create a lesson",
                "operationId": "createLesson",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id",
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "string",
                                        "example": "665c9f2e8f1b2a4d7c123457"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Routing Basics"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Learn how Laravel routes requests."
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lessons/{lesson}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get a lesson",
                "operationId": "getLesson",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "MongoDB lesson ObjectId.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "665c9f2e8f1b2a4d7c123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Lesson not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update a lesson",
                "operationId": "updateLesson",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "MongoDB lesson ObjectId.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "665c9f2e8f1b2a4d7c123456"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "string",
                                        "example": "665c9f2e8f1b2a4d7c123457"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Routing Basics"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Learn how Laravel routes requests."
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Lesson not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete a lesson",
                "operationId": "deleteLesson",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "MongoDB lesson ObjectId.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "665c9f2e8f1b2a4d7c123456"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lesson deleted successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Lesson not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Lesson": {
                "required": [
                    "id",
                    "course_id",
                    "title",
                    "content"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true,
                        "example": "665c9f2e8f1b2a4d7c123456"
                    },
                    "course_id": {
                        "type": "string",
                        "example": "665c9f2e8f1b2a4d7c123457"
                    },
                    "title": {
                        "type": "string",
                        "example": "Routing Basics"
                    },
                    "content": {
                        "type": "string",
                        "example": "Learn how Laravel routes requests."
                    },
                    "position": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}